Working on circulation workflow, Roles and permissiions and Interfaces

+++ app/Http/Livewire/MailsTable.php
@@ -2,6 +2,10 @@
 
 namespace App\Http\Livewire;
 
+use App\Models\Register;
+use App\Models\Scopes\MailScope;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Laracasts\Flash\Flash;
 use Rappasoft\LaravelLivewireTables\DataTableComponent;
 use Rappasoft\LaravelLivewireTables\Views\Column;
@@ -9,12 +13,15 @@ use App\Models\Mail;
 
 class MailsTable extends DataTableComponent
 {
+    use AuthorizesRequests;
+
     protected $model = Mail::class;
 
     protected $listeners = ['deleteRecord' => 'deleteRecord'];
 
     public function deleteRecord($id)
     {
+        $this->authorize('ViewAny', Mail::class);
         Mail::find($id)->delete();
         Flash::success(__('messages.deleted', ['model' => __('models/mails.singular')]));
         $this->emit('refreshDatatable');
@@ -25,9 +32,18 @@ class MailsTable extends DataTableComponent
         $this->setPrimaryKey('id');
     }
 
+    public function builder(): Builder
+    {
+        return $this->getModel()::query()->with($this->getRelationships())->withoutGlobalScope(MailScope::class);
+    }
+
     public function columns(): array
     {
         return [
+            Column::make("Id", "id")
+                ->sortable()
+                ->searchable(),
+
             Column::make("Objet", "objet")
                 ->sortable()
                 ->searchable(),
@@ -44,6 +60,7 @@ class MailsTable extends DataTableComponent
                         'recordUrl' => route('circulation.record', $row->id),
                         'sendUrl' => route('circulation.send', $row->id),
                         'processingUrl' => route('circulation.processing', $row->id),
+                        'SendProcessingUrl' => route('circulation.SendProcessing', $row->id),
                         'editUrl' => route('mails.edit', $row->id),
                         'recordId' => $row->id,
                         'title' => $row->objet,
